server test: refactoring
