+++ src/tests/tools/iotc_test_echoserver.cc
@@ -70,64 +70,16 @@ uint16_t EchoTestServer::RunServer() {
       return 1;
     }
   } else if (sock_type_ == SOCK_DGRAM) {
-    if (RunTdpServer() != 0) {
+    if (RunUdpServer() != 0) {
       return 1;
     }
   }
-  return 0;
-}
-
-uint16_t EchoTestServer::RunTcpServer() {
-  struct sockaddr_storage client_addr;
-  socklen_t client_addr_size = sizeof(struct sockaddr_storage);
-
-  /* Accept and read message from client. */
-  while (runnable_) {
-    client_sock_ =
-        accept(server_sock_, (struct sockaddr*)&client_addr, &client_addr_size);
-    if (client_sock_ < 0) {
-      close(client_sock_);
-      close(server_sock_);
-      return 1;
-    }
-    recv_len_ = read(client_sock_, recv_buf_, kBufferSize);
-    recv_buf_[recv_len_] = '\0';
-    write(client_sock_, recv_buf_, recv_len_);
-    close(client_sock_);
-  }
-
-  close(server_sock_);
-  return 0;
-}
-
-uint16_t EchoTestServer::RunTdpServer() {
-  struct sockaddr_storage client_addr;
-  socklen_t client_addr_size = sizeof(struct sockaddr_storage);
-
-  /* Receive message from client. */
-  while (runnable_) {
-    if ((recv_len_ =
-             recvfrom(server_sock_, recv_buf_, kBufferSize, 0,
-                      (struct sockaddr*)&client_addr, &client_addr_size)) < 0) {
-      close(server_sock_);
-      return 1;
-    }
-    recv_buf_[recv_len_] = '\0';
-    if (sendto(server_sock_, recv_buf_, recv_len_, 0,
-               (struct sockaddr*)&client_addr, client_addr_size) != recv_len_) {
-      perror("[Server] Error sending to client");
-      close(server_sock_);
-      return 1;
-    }
-  }
-
-  close(server_sock_);
   return 0;
 }
 
 char* EchoTestServer::get_recv_buf() { return recv_buf_; }
 
-void EchoTestServer::stop_server() {
+void EchoTestServer::StopServer() {
   runnable_ = false;
   return;
 }
